support python3 and python2
python3: build with "--tags=py3.5"
python2: default, without build tag
