+++ cgolinks.go
@@ -1,6 +1,10 @@
+// +build !py3.4
+// +build !py3.5
+// +build !py3.6
+
 package py
 
 /*
-#cgo pkg-config: python-3.5
+#cgo pkg-config: python-2.7
 */
 import "C"
