+++ cgolinks_py3.5.go
@@ -1,0 +1,8 @@
+// +build py3.5
+
+package py
+
+/*
+#cgo pkg-config: python-3.5
+*/
+import "C"
