+++ error_py2_test.go
@@ -1,0 +1,9 @@
+// +build !py3.4
+// +build !py3.5
+// +build !py3.6
+
+package py
+
+func loadExceptionModule() (ObjectModule, error) {
+	return LoadModule("exceptions")
+}
