+++ mainthread/cgolinks.go
@@ -1,0 +1,10 @@
+// !py3.4
+// !py3.5
+// !py3.6
+
+package mainthread
+
+/*
+#cgo pkg-config: python-2.7
+*/
+import "C"
