+++ mainthread/cgolinks_py3.4.go
@@ -1,0 +1,8 @@
+// +build py3.4
+
+package mainthread
+
+/*
+#cgo pkg-config: python-3.4
+*/
+import "C"
